Updated README with justification and better summary
Notice I deleted the list of steps, but kept what was more relevant from them into the Scripts documentation section. Also, I moved the affiliations list to the top, to distinguish them from the reference list. Reggarding the refernece list, the second time the same reference is mentioned in the text it doesn't appear as a supscript and link, but just as plain text incluiding the "^". Not sure how to fix it, maybe using normal formmating the numbers as supcript without the ference link?